requirements_v2.txt
# ============================================================
# SWARNE V2.0 - REQUIREMENTS
# Sprint de développement 24h - SYSTÈME COMPLET
# ============================================================

# ============================================================
# CORE DEPENDENCIES (V1)
# ============================================================
numpy>=1.21.0
pandas>=1.3.0
MetaTrader5>=5.0.37
scipy>=1.7.0

# ============================================================
# MACHINE LEARNING & AI
# ============================================================
# Deep Learning
tensorflow>=2.13.0              # LSTM, Transformers
torch>=2.0.0                    # PyTorch pour modèles alternatifs
scikit-learn>=1.3.0             # ML classique
keras>=2.13.0                   # High-level API

# Feature Engineering
ta>=0.10.0                      # Technical Analysis indicators
ta-lib>=0.4.25                  # TA-Lib (avancé)
pandas-ta>=0.3.14b              # pandas Technical Analysis

# Optimization
optuna>=3.3.0                   # Hyperparameter optimization
ray[tune]>=2.7.0                # Distributed training

# ============================================================
# BACKTESTING & ANALYTICS
# ============================================================
backtrader>=1.9.78              # Backtesting framework (alternative)
vectorbt>=0.25.0                # Vectorized backtesting
pyfolio>=0.9.2                  # Performance analytics
empyrical>=0.5.5                # Financial metrics

# ============================================================
# VISUALIZATION & DASHBOARD
# ============================================================
PyQt5>=5.15.9                   # GUI framework
PyQtChart>=5.15.6               # Charts pour PyQt5
pyqtgraph>=0.13.3               # Real-time plotting
matplotlib>=3.7.0               # Plotting
seaborn>=0.12.0                 # Statistical visualization
plotly>=5.17.0                  # Interactive plots

# ============================================================
# DATABASE & STORAGE
# ============================================================
SQLAlchemy>=2.0.0               # ORM
alembic>=1.12.0                 # Database migrations
psycopg2-binary>=2.9.7          # PostgreSQL (optionnel)
pymongo>=4.5.0                  # MongoDB (optionnel)
redis>=5.0.0                    # Cache

# ============================================================
# NOTIFICATIONS & COMMUNICATION
# ============================================================
python-telegram-bot>=20.5       # Telegram bot
twilio>=8.9.0                   # SMS notifications
sendgrid>=6.10.0                # Email notifications
requests>=2.31.0                # HTTP requests

# ============================================================
# DATA HANDLING
# ============================================================
yfinance>=0.2.28                # Yahoo Finance data
alpha-vantage>=2.3.1            # Alpha Vantage API
ccxt>=4.1.0                     # Crypto exchanges
pandas-datareader>=0.10.0       # Data readers

# ============================================================
# OPTIMIZATION & PERFORMANCE
# ============================================================
numba>=0.58.0                   # JIT compilation
cython>=3.0.2                   # C extensions
joblib>=1.3.0                   # Parallel processing
dask>=2023.9.0                  # Parallel computing

# ============================================================
# CONFIGURATION & UTILITIES
# ============================================================
pyyaml>=6.0.1                   # YAML config
python-dotenv>=1.0.0            # Environment variables
click>=8.1.7                    # CLI interface
rich>=13.5.0                    # Rich terminal output
tqdm>=4.66.0                    # Progress bars
loguru>=0.7.0                   # Advanced logging

# ============================================================
# TESTING & QUALITY
# ============================================================
pytest>=7.4.0                   # Testing framework
pytest-cov>=4.1.0               # Coverage
pytest-asyncio>=0.21.0          # Async testing
black>=23.7.0                   # Code formatting
flake8>=6.1.0                   # Linting
mypy>=1.5.0                     # Type checking

# ============================================================
# WEB & API (BONUS)
# ============================================================
fastapi>=0.103.0                # REST API
uvicorn>=0.23.0                 # ASGI server
websockets>=11.0.0              # WebSocket support
streamlit>=1.27.0               # Web dashboard alternative

# ============================================================
# CRYPTO & SECURITY
# ============================================================
cryptography>=41.0.0            # Encryption
pyjwt>=2.8.0                    # JWT tokens

# ============================================================
# DEVELOPMENT TOOLS
# ============================================================
ipython>=8.15.0                 # Enhanced Python shell
jupyter>=1.0.0                  # Notebooks
jupyterlab>=4.0.0               # Lab interface

# ============================================================
# OPTIONAL: GPU ACCELERATION
# ============================================================
# tensorflow-gpu>=2.13.0        # TensorFlow with GPU
# torch-cuda>=2.0.0             # PyTorch with CUDA

# ============================================================
# NOTES:
# - Installer avec: pip install -r requirements_v2.txt
# - Pour GPU: Installer CUDA Toolkit 11.8+ séparément
# - Ta-lib nécessite installation système: apt-get install ta-lib
# ============================================================
